+++ client/ui/Credits.jsx
@@ -32,8 +32,14 @@ export default class Credits extends Component {
                       <h5><b>Release information</b></h5>
 
                       <ul>
+                        <b>6.9: </b>
+                        <li>Added function to hide unmatching layers when searching across all layers</li>
+                        <li>Feature added to click on the translation to disable all others from search</li>
+                      </ul>
+
+                      <ul>
                         <b>6.8 </b>
-                        <li>Commentaries and other verse details loading from ES in the iFrame</li>                        
+                        <li>Commentaries and other verse details loading from ES in the iFrame</li>
                       </ul>
 
                       <ul>