+++ client/ui/Results.jsx
@@ -28,7 +28,9 @@ class Results extends Component {
                       setVerse={this.setVerse.bind(this)}
                       options={this.props.options}
                       search={this.props.search.bind(this)}
-                      analyzers={this.props.analyzers}/>
+                      hideUnmatched={this.props.hideUnmatched}
+                      analyzers={this.props.analyzers}
+                      switchLayers={this.props.switchLayers}/>
 
                ))}
                <Paging total={r.results.hits.total.value}
@@ -37,8 +39,8 @@ class Results extends Component {
                   search={this.props.search.bind(this)}
                   page={this.props.page}
                   limit={this.props.limit} />
-              </div>:<div className="NoResult"><h4>Sorry, no results found</h4><hr/><h5>If like to perform a global search instead, click
-                      <a onClick={this.globalSearch.bind(this, 1, this.props.limit)}> here</a><br/><br/>See advance filter examples below for more details</h5><hr/><Help/></div>:"...":<Help/>
+              </div>:<div className="NoResult"><h4>Sorry, no results found</h4><hr/><h5>If like to perform a global search instead,
+                click <a onClick={this.globalSearch.bind(this, 1, this.props.limit)}>here</a><br/><br/>See advance filter examples below for more details</h5><hr/><Help/></div>:"...":<Help/>
        }
        </div>
  }
@@ -49,12 +51,20 @@ class Results extends Component {
  setPage(page,limit) {
     this.props.setPage(page,limit);
  }
-
+//{this.globalSearch(this, 1, this.props.limit)}
  globalSearch(page,limit) {
-   window.query = ":\""+window.query.replace(/"|\:/g,'') + "\""
+   // console.log(page,limit);
+   // window.query = ":\""+window.query.replace(/"|\:/g,'') + "\""
    $(window.inputId)[0].value = window.query
-   // this.props.options.map(y=>{y.state=true})
-   this.props.search(window.query,this.props.options,page,limit)
+   this.props.switchLayers()
+   this.props.options.map(y=>{if (!y.id.match(/Chinese/i)){y.state=true}})
+
+   try {
+     // setTimeout(this.props.search, 1000, window.query, this.props.options,page,limit);
+     this.props.search(window.query,this.props.options,page,limit)
+   } catch (e) {
+
+   }
 
    // alert(1)
  }
@@ -62,7 +72,7 @@ class Results extends Component {
 }
 
 Results.propTypes = {
- //results: PropTypes.object.isRequired
+ // results: PropTypes.object.isRequired
 }
 
 export default createContainer(props => {